fix(ui) Attempt to reduce errors from autoplaying videos
The changes to loadingIndicator have worked pretty well. I wanted to try
the same approach for the CongratsRobot to reduce errors more.

Fixes JAVASCRIPT-21B4
