+++ src/sentry/static/sentry/app/views/issueList/noGroupsHandler/congratsRobots.tsx
@@ -3,18 +3,19 @@ import styled from '@emotion/styled';
 
 import space from 'app/styles/space';
 import video from 'app/../images/spot/congrats-robots.mp4';
+import AutoplayVideo from 'app/components/autoplayVideo';
 
 /**
  * Note, video needs `muted` for `autoplay` to work on Chrome
  * See https://developer.mozilla.org/en-US/docs/Web/HTML/Element/video
  */
-const CongratsRobots = () => (
-  <AnimatedScene>
-    <StyledVideo autoPlay loop muted>
-      <source src={video} type="video/mp4" />
-    </StyledVideo>
-  </AnimatedScene>
-);
+function CongratsRobots() {
+  return (
+    <AnimatedScene>
+      <StyledAutoplayVideo src={video} />
+    </AnimatedScene>
+  );
+}
 
 export default CongratsRobots;
 
@@ -22,7 +23,7 @@ const AnimatedScene = styled('div')`
   max-width: 800px;
 `;
 
-const StyledVideo = styled('video')`
+const StyledAutoplayVideo = styled(AutoplayVideo)`
   max-height: 320px;
   max-width: 100%;
   margin-bottom: ${space(1)};
